Change type of n_descriptors to avoid an unnecessary cast

+++ source/descriptor_heap.cpp
@@ -7,7 +7,7 @@
 #include "device.h"
 
 namespace gfx {
-    DescriptorHeap::DescriptorHeap(const Device& device, D3D12_DESCRIPTOR_HEAP_TYPE type, D3D12_DESCRIPTOR_HEAP_FLAGS flags, size_t n_descriptors) {
+    DescriptorHeap::DescriptorHeap(const Device& device, D3D12_DESCRIPTOR_HEAP_TYPE type, D3D12_DESCRIPTOR_HEAP_FLAGS flags, uint32_t n_descriptors) {
         const D3D12_DESCRIPTOR_HEAP_DESC desc = {
             .Type = type,
             .NumDescriptors = static_cast<UINT>(n_descriptors * 2),
